services/chat-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
firebase-admin==6.2.0
python-jose==3.3.0
httpx==0.24.1
python-multipart==0.0.6
starlette==0.27.0
pydantic-settings==2.0.3
google-cloud-firestore==2.11.1 